+++ HISTORY.rst
@@ -21,9 +21,15 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* Stock picking scrap: update next picking qty BSMTS-197
+
+
 **Bugfixes**
 
 **Build**
+
+* Remove dummy package
+
 
 **Documentation**
 
@@ -47,8 +53,6 @@ latest (unreleased)
 * There is no longer default value for comment in 
   invoice BSMTS-216
 * Add 'tax display' field in sale settings is now B2C BSMTS-217
-* Scraping in stock.picking now refreshes the status and
-  units BSMTS-197
 
 **Bugfixes**