+++ HISTORY.rst
@@ -26,6 +26,7 @@ latest (unreleased)
 * Add changes to py3o reports BSMTS-223
 * Add project analysis wqeb report changes BSMTS-214
 * Add language support to the chemical analysis BSMTS-221
+* Renamed conformity field to compliance in product.substance.measure BSMTS-220
 
 **Bugfixes**